master

+++ app/src/main/java/my/neomer/tapper/IPhysicsObject.java
@@ -5,6 +5,6 @@ public interface IPhysicsObject {
     void ApplyForce(Vector force);
     void ApplyImpulse(Vector impulse);
 
-    void UpdatePhysics(long timespan);
+    void UpdatePhysics(double timespan);
 
 }
